requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
psycopg[binary]>=3.1.0
pydantic>=2.6.0
python-dotenv>=1.0.0
reportlab>=4.0.0